+++ Changes.md
@@ -1,5 +1,6 @@
 Release 16.1.0
 ==============
+* Import awards script information into a new database
 * #966 - Remove edit score when entering performance scores for a single table 
 * #964 - Add report of subjective scores by judge
 * #942 - improve the visibility of the left and right portions of a condition